+++ lib/app.dart
@@ -24,7 +24,7 @@
 export 'package:mvc_pattern/mvc_pattern.dart' show AppMVC, AppConMVC;
 
 export 'package:mvc_application/src/controller/app.dart'
-    show AppController, ErrorHandler;
+    show AppController, ErrorHandler, runApp;
 
 export 'package:mvc_application/src/view/app.dart' show App, AppDrawer;