+++ lib/src/view/uxutils/src/view/DialogBox.dart
@@ -19,9 +19,13 @@
 //
 //          Created  05 Feb 2019
 //
-//
+// DialogBox.dart
 
 import 'package:flutter/material.dart';
+
+import 'package:flutter/cupertino.dart';
+
+import 'package:mvc_application/view.dart';
 
 Future<bool> showBox({
   @required BuildContext context,
@@ -39,28 +43,51 @@ Future<bool> showBox({
   }
   final ThemeData theme = Theme.of(context);
   final TextStyle dialogTextStyle =
-      theme.textTheme.subhead.copyWith(color: theme.textTheme.caption.color);
-  return showDialog<bool>(
-    context: context,
-    builder: (BuildContext context) => AlertDialog(
-        content: Text(text ?? ' ', style: dialogTextStyle),
-        actions: <Widget>[
-          FlatButton(
-              child: Text(button02.text ?? 'Cancel'),
-              onPressed: () {
-                if (press02 != null) press02();
-                if (button02.onPressed != null) button02.onPressed();
-                Navigator.pop(context, button02.result);
-              }),
-          FlatButton(
-              child: Text(button01.text ?? 'OK'),
-              onPressed: () {
-                if (press01 != null) press01();
-                if (button01.onPressed != null) button01.onPressed();
-                Navigator.pop(context, button01.result);
-              }),
-        ]),
-  );
+      theme.textTheme.subtitle1.copyWith(color: theme.textTheme.caption.color);
+  if (App.useMaterial)
+    return showDialog<bool>(
+      context: context,
+      builder: (BuildContext context) => AlertDialog(
+          content: Text(text ?? ' ', style: dialogTextStyle),
+          actions: <Widget>[
+            FlatButton(
+                child: Text(button02.text ?? 'Cancel'),
+                onPressed: () {
+                  if (press02 != null) press02();
+                  if (button02.onPressed != null) button02.onPressed();
+                  Navigator.pop(context, button02.result);
+                }),
+            FlatButton(
+                child: Text(button01.text ?? 'OK'),
+                onPressed: () {
+                  if (press01 != null) press01();
+                  if (button01.onPressed != null) button01.onPressed();
+                  Navigator.pop(context, button01.result);
+                }),
+          ]),
+    );
+  if (App.useCupertino)
+    return showCupertinoDialog<bool>(
+      context: context,
+      builder: (BuildContext context) => CupertinoAlertDialog(
+          content: Text(text ?? ' ', style: dialogTextStyle),
+          actions: <Widget>[
+            CupertinoDialogAction(
+                child: Text(button02.text ?? 'Cancel'),
+                onPressed: () {
+                  if (press02 != null) press02();
+                  if (button02.onPressed != null) button02.onPressed();
+                  Navigator.pop(context, button02.result);
+                }),
+            CupertinoDialogAction(
+                child: Text(button01.text ?? 'OK'),
+                onPressed: () {
+                  if (press01 != null) press01();
+                  if (button01.onPressed != null) button01.onPressed();
+                  Navigator.pop(context, button01.result);
+                }),
+          ]),
+    );
 }
 
 void dialogBox({