+++ lib/view.dart
@@ -31,6 +31,10 @@ export 'package:mvc_pattern/mvc_pattern.dart' show AppMVC, ViewMVC, SetState;
 /// Settings
 export 'package:mvc_application/src/view/utils/appsettings.dart';
 
+/// Error Handling
+export 'package:mvc_application/src/view/utils/error_handler.dart'
+    show ErrorHandler, ReportErrorHandler;
+
 /// Screens
 export 'package:mvc_application/src/view/utils/loading_screen.dart';